Add streaming_run_events with examples on how to use new streaming methods
