+++ autogen/agentchat/__init__.py
@@ -21,6 +21,7 @@ from .contrib.swarm_agent import (
     SwarmResult,
     a_initiate_swarm_chat,
     initiate_swarm_chat,
+    initiate_swarm_chat_stream,
     register_hand_off,
 )
 from .conversable_agent import UPDATE_SYSTEM_MESSAGE, ConversableAgent, UpdateSystemMessage, register_function
@@ -54,6 +55,7 @@ __all__ = [
     "gather_usage_summary",
     "initiate_chats",
     "initiate_swarm_chat",
+    "initiate_swarm_chat_stream",
     "register_function",
     "register_hand_off",
 ]
