+++ autogen/io/run_response.py
@@ -10,7 +10,7 @@ from uuid import UUID, uuid4
 
 from ..events.agent_events import ErrorEvent, InputRequestEvent, TerminationEvent
 from ..events.base_event import BaseEvent
-from ..events.print_event import PrintEvent
+from .thread_io_stream import ThreadIOStream
 
 Message = dict[str, Any]
 
@@ -43,27 +43,6 @@ class AsyncRunResponseProtocol(RunInfoProtocol, Protocol):
 
     @property
     async def summary(self) -> str: ...
-
-
-class ThreadIOStream:
-    def __init__(self) -> None:
-        self._input_stream: queue.Queue = queue.Queue()  # type: ignore[type-arg]
-        self._output_stream: queue.Queue = queue.Queue()  # type: ignore[type-arg]
-
-    def input(self, prompt: str = "", *, password: bool = False) -> str:
-        self.send(InputRequestEvent(prompt=prompt, password=password))  # type: ignore[call-arg]
-        return self._output_stream.get()  # type: ignore[no-any-return]
-
-    def print(self, *objects: Any, sep: str = " ", end: str = "\n", flush: bool = False) -> None:
-        print_message = PrintEvent(*objects, sep=sep, end=end)
-        self.send(print_message)
-
-    def send(self, message: Any) -> None:
-        self._input_stream.put(message)
-
-    @property
-    def input_stream(self) -> queue.Queue:  # type: ignore[type-arg]
-        return self._input_stream
 
 
 class RunResponse:
